Isolate Duration parsing workaround and reduce allocations
Move Duration parsing code into a separate object and reduce allocations by removing Either and throwing exceptions (sorry!)
